F09/F10 - refactor: consolidate session type imports and remove unused types from dashboard
