--- src/lib/server/db/types.ts
@@ -1,3 +1,0 @@
-import type { session } from './schema';
-
-export type Session = typeof session.$inferSelect;
