+++ src/lib/types.ts
@@ -1,8 +1,10 @@
 import type { Component } from 'svelte';
 
-import type { user } from './server/db/schema';
+import type { session, user } from './server/db/schema';
 
 export type User = typeof user.$inferSelect;
+
+export type Session = typeof session.$inferSelect;
 
 export type SidebarNav = {
 	navMain: {
