+++ src/routes/(app)/dashboard/+page.server.ts
@@ -18,10 +18,7 @@ import {
 	formatTimestampMedium,
 	formatTimestampShort,
 	formatTodayLabel,
-	getISODayOfWeek,
-	getStartOfWeek,
-	getTodayString,
-	getWeekDates
+	getTodayString
 } from '$lib/utils/date';
 import { logger } from '$lib/utils/logger';
 import { createMarkdownExcerpt } from '$lib/utils/markdown';
@@ -30,6 +27,7 @@ import { getWorkoutLabel } from '$lib/utils/workout';
 import { and, desc, eq, gte, lte, ne, sql } from 'drizzle-orm';
 
 import type { PageServerLoad } from './$types';
+import { buildAgendaItemStats, buildDateRanges, type DashboardDateRanges } from './dashboard-utils';
 
 type ActivityItem = {
 	type: 'journal' | 'workout' | 'meditation' | 'task' | 'visit';
@@ -45,32 +43,6 @@ type VisitHealthBucket = 'critical' | 'overdue' | 'healthy' | 'noVisits';
 type VisitHealthResult = {
 	counts: Record<VisitHealthBucket, number> & { total: number };
 	names: Record<VisitHealthBucket, string[]>;
-};
-
-type DashboardDateRanges = {
-	today: string;
-	startOfThisWeekDate: string;
-	endOfThisWeekDate: string;
-	startOfLastWeekDate: string;
-	startOfMonthDate: string;
-	thisWeekDates: string[];
-	lastWeekDates: string[];
-	meditationRangeStartIso: string;
-	meditationRangeEndIso: string;
-	agenda8WeeksStart: string;
-	last4WeeksStart: string;
-	workout4WeeksStart: string;
-	upcomingEndDate: string;
-	tasksCompletionRangeStartIso: string;
-	tasksCompletionRangeEndIso: string;
-};
-
-type ItemBucket = {
-	title: string;
-	titleDate: string;
-	last4: { total: number; completed: number };
-	prev4: { total: number; completed: number };
-	dowLast4: { total: number[]; completed: number[] };
 };
 
 const STATUS_KEY_MAP: Record<string, VisitHealthBucket> = {
@@ -158,32 +130,6 @@ function buildVisitHealth(
 	}
 	return { counts, names };
 }
-
-// ── Date ranges ────────────────────────────────────────────────────────────────
-function buildDateRanges(today: string): DashboardDateRanges {
-	const startOfThisWeekDate = getStartOfWeek(today);
-	const endOfThisWeekDate = addDaysToDateString(startOfThisWeekDate, 7);
-	const startOfLastWeekDate = addDaysToDateString(startOfThisWeekDate, -7);
-	return {
-		today,
-		startOfThisWeekDate,
-		endOfThisWeekDate,
-		startOfLastWeekDate,
-		startOfMonthDate: addDaysToDateString(today, -30),
-		thisWeekDates: getWeekDates(startOfThisWeekDate),
-		lastWeekDates: getWeekDates(startOfLastWeekDate),
-		meditationRangeStartIso: `${addDaysToDateString(startOfLastWeekDate, -1)}T00:00:00.000Z`,
-		meditationRangeEndIso: `${addDaysToDateString(endOfThisWeekDate, 1)}T00:00:00.000Z`,
-		agenda8WeeksStart: addDaysToDateString(startOfThisWeekDate, -49),
-		last4WeeksStart: addDaysToDateString(startOfThisWeekDate, -21),
-		workout4WeeksStart: addDaysToDateString(today, -28),
-		upcomingEndDate: addDaysToDateString(today, 7),
-		tasksCompletionRangeStartIso: `${addDaysToDateString(startOfLastWeekDate, -1)}T00:00:00.000Z`,
-		tasksCompletionRangeEndIso: `${addDaysToDateString(endOfThisWeekDate, 1)}T00:00:00.000Z`
-	};
-}
-
-export const _buildDateRanges = buildDateRanges;
 
 // ── Database queries ───────────────────────────────────────────────────────────
 async function runDashboardQueries(
@@ -392,17 +338,6 @@ async function runDashboardQueries(
 
 type QueryData = Awaited<ReturnType<typeof runDashboardQueries>>;
 
-type AgendaItemStat = {
-	title: string;
-	completionPct: number;
-	prevCompletionPct: number;
-	dowCompletionPct: number[];
-	last4Days: number;
-};
-
-type AgendaRangeWindow = 'last4' | 'prev4';
-type AgendaTrendEntry = QueryData['agendaEntriesForTrend'][number];
-
 // ── Post-processing helpers ────────────────────────────────────────────────────
 function buildCountByDateMap(rows: { date: string; count: number }[]): Map<string, number> {
 	return new Map(rows.map((r) => [r.date, Number(r.count || 0)]));
@@ -536,110 +471,6 @@ function buildActivityFeed(
 		.sort((a, b) => b.timestamp.localeCompare(a.timestamp))
 		.slice(0, 10);
 }
-
-function getAgendaRangeWindow(
-	date: string,
-	ranges: Pick<DashboardDateRanges, 'agenda8WeeksStart' | 'last4WeeksStart' | 'endOfThisWeekDate'>
-): AgendaRangeWindow | null {
-	if (date >= ranges.last4WeeksStart && date < ranges.endOfThisWeekDate) {
-		return 'last4';
-	}
-
-	if (date >= ranges.agenda8WeeksStart && date < ranges.last4WeeksStart) {
-		return 'prev4';
-	}
-
-	return null;
-}
-
-function createItemBucket(entry: AgendaTrendEntry): ItemBucket {
-	return {
-		title: entry.title,
-		titleDate: entry.date,
-		last4: { total: 0, completed: 0 },
-		prev4: { total: 0, completed: 0 },
-		dowLast4: {
-			total: Array.from({ length: 7 }, () => 0),
-			completed: Array.from({ length: 7 }, () => 0)
-		}
-	};
-}
-
-function getOrCreateItemBucket(
-	itemBuckets: Map<string, ItemBucket>,
-	groupKey: string,
-	entry: AgendaTrendEntry
-): ItemBucket {
-	let bucket = itemBuckets.get(groupKey);
-	if (!bucket) {
-		bucket = createItemBucket(entry);
-		itemBuckets.set(groupKey, bucket);
-	}
-	return bucket;
-}
-
-function addEntryToBucket(
-	bucket: ItemBucket,
-	entry: AgendaTrendEntry,
-	window: AgendaRangeWindow
-): void {
-	if (entry.date > bucket.titleDate) {
-		bucket.title = entry.title;
-		bucket.titleDate = entry.date;
-	}
-
-	const period = window === 'last4' ? bucket.last4 : bucket.prev4;
-	period.total++;
-	if (entry.completed) period.completed++;
-
-	if (window === 'last4') {
-		const dowIndex = getISODayOfWeek(entry.date);
-		bucket.dowLast4.total[dowIndex]++;
-		if (entry.completed) bucket.dowLast4.completed[dowIndex]++;
-	}
-}
-
-function toPct(completed: number, total: number): number {
-	return total > 0 ? Math.round((completed / total) * 100) : 0;
-}
-
-function toDowCompletionPct(bucket: ItemBucket): number[] {
-	return bucket.dowLast4.total.map((tot, i) =>
-		tot > 0 ? Math.round((bucket.dowLast4.completed[i] / tot) * 100) : -1
-	);
-}
-
-function toAgendaItemStat(bucket: ItemBucket): AgendaItemStat {
-	return {
-		title: bucket.title,
-		completionPct: toPct(bucket.last4.completed, bucket.last4.total),
-		prevCompletionPct: toPct(bucket.prev4.completed, bucket.prev4.total),
-		dowCompletionPct: toDowCompletionPct(bucket),
-		last4Days: bucket.last4.total
-	};
-}
-
-function buildAgendaItemStats(
-	entries: QueryData['agendaEntriesForTrend'],
-	ranges: DashboardDateRanges
-): AgendaItemStat[] {
-	const itemBuckets = new Map<string, ItemBucket>();
-
-	for (const entry of entries) {
-		const window = getAgendaRangeWindow(entry.date, ranges);
-		if (!window) continue;
-		const groupKey = entry.templateGroupId ?? entry.title;
-		const bucket = getOrCreateItemBucket(itemBuckets, groupKey, entry);
-		addEntryToBucket(bucket, entry, window);
-	}
-
-	return Array.from(itemBuckets.values())
-		.filter((b) => b.last4.total + b.prev4.total >= 7)
-		.map(toAgendaItemStat)
-		.sort((a, b) => a.completionPct - b.completionPct);
-}
-
-export const _buildAgendaItemStats = buildAgendaItemStats;
 
 function buildTodayAgendaSummary(entries: QueryData['agendaEntriesForTrend'], today: string) {
 	const todayEntries = entries
